requirements.txt
asgiref==3.6.0
astroid==2.15.2
dill==0.3.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
Faker==18.4.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
pylint==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
typed-ast==1.5.4
typing_extensions==4.5.0
validate-docbr==1.10.0
wrapt==1.12.1
